refactor(core,linear,geometry): dedup shared math primitives into core::linalg (C1-FOLLOWUP)
The C1 port (PR #62) deliberately left `vision-calibration-linear` untouched,
duplicating the low-level numeric primitives that also live in the fresh
`vision-geometry` crate. This collapses that duplication for the `math` module —
the cleanest, numerically-safe slice.

The primitives `linear` and `geometry` share are pure linear-algebra helpers
(Hartley normalization, polynomial roots, homogeneous null-space, SVD/vector
reshaping) — not geometry-specific. Rather than make the *published*
`vision-calibration-linear` depend on the (private) `vision-geometry` for them — a
layering smell that would also force geometry's premature publication — they move
*down* into the shared foundation crate as a new `vision_calibration_core::linalg`
module. Both `linear::math` and `geometry::math` now re-export from it.

- `vision_calibration_core::linalg` is the canonical home: `normalize_points_2d/3d`,
  `solve_{quadratic,cubic,quartic}_real`, `null_space` + `NullSpaceSolution`,
  `mat3/mat34_{from_vec,from_svd_row}`. Verified logically identical to both
  prior copies (9/10 byte-identical; `null_space` differed only in error wrapping).
- A typed `MathError` (in `core::linalg`) replaces stringly `anyhow` for
  `null_space`, advancing Track D / D1. Geometry's internal anyhow callers still
  convert via `?`; `linear` maps it to `Error::Singular` via `From`.
- `geometry::math` keeps only its DLT-specific `dlt_rank_ok`; `linear` keeps the
  calibration-only `ridge_lstsq` / `project_to_so3`.
- No new cross-crate dependency edge (no `linear → geometry`), so geometry's
  crates.io publication stays a separate, deliberate decision.

Full workspace suite green (671 tests, zero numeric drift); fmt + clippy +
rustdoc(-D warnings) clean.

The divergent higher-level solvers (homography, epipolar, camera_matrix,
triangulation) remain duplicated between `linear` and `geometry` and are deferred
to a supervised follow-up (numeric-equivalence reconciliation needed first) — see
docs/backlog.md C1-FOLLOWUP.
